+++ src/FeatureIntro/index.ts
@@ -1,0 +1,6 @@
+export { FeatureIntroControl as Control } from "../FeatureIntroControl";
+export { FeatureIntroContent as Content } from "../FeatureIntroContent";
+export { FeatureIntroHeading as Heading } from "../FeatureIntroHeading";
+export { FeatureIntroImage as Image } from "../FeatureIntroImage";
+export { FeatureIntroLearnMoreLink as LearnMoreLink } from "../FeatureIntroLearnMoreLink";
+export { FeatureIntroDismissButton as DismissButton } from "../FeatureIntroDismissButton";